Actuators
A tinny led have been added to the proyect, representing the multiple
ways to send processed info by Arduino microcontroller.

This led also guarantees the board is processing the info and taking
decisions to act. In particular it does emmit different intensity
depending on (and proportional) difference between temperature
measure and chosen limit by user.

A proper design should include a wireless comm module, to send that info
to another board. However, this part of the project will extend it too
much. By time reasons we won't implement this part at this moment.