Adding useful examples and updating title
A few useful examples have been added. Those examples have been used to
calibrate sensors, try features or general testing. They are
EXPERIMENTAL code, some comments could not be in english but in general
it is a much more simple code than src, because they are used to
isolated modules.